backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
PyMuPDF==1.24.14
python-docx==1.1.2
google-generativeai==0.8.3
python-dotenv==1.0.1
aiofiles==24.1.0
psycopg2-binary==2.9.10
celery[redis]==5.4.0
PyJWT==2.9.0
bcrypt==4.2.0
slowapi==0.1.9
